refactor(cli): split model registry and session listing; fix log rotation race
- models/model-spec.ts: registry types, spec parsing/resolution/formatting
  and applyProvider move out of model-registry.ts, which keeps only the
  load/merge/normalize half. Importers now import each half explicitly
  (no re-export shim). 162 + 200 lines.
- commands/session-listing.ts: the pure print halves of /sessions,
  /search and the resume-time recent-conversation preview move out of
  SessionCommands (263 + 75 lines); output is byte-identical.
- shared/log-sink.ts: EngineLogSink.restore() now awaits an in-flight
  rotation before flushing. rotate()'s rename+reopen ride on an async
  end() flush; under load restore() could resolve mid-swap and the
  uninstalled guard then skipped the rename, losing the .old swap the
  caller was told had settled — observed as a ~50% failure of the
  rotation spec under full-suite parallel load. rotationSettled makes
  the swap awaitable; the existing rotation test now passes 4/4 full
  runs.
